+++ src/main/java/seul/bowling/domain/pin/Pin.java
@@ -1,4 +1,4 @@
-package seul.bowling.domain;
+package seul.bowling.domain.pin;
 
 import lombok.Getter;
 import seul.bowling.exception.BowlingException;
@@ -10,11 +10,15 @@ public class Pin {
     @Getter
     private int pin;
 
-    public Pin(int pin) {
+    private Pin(int pin) {
         if (pin > MAX_PIN_COUNT) {
             throw new BowlingException(ExceptionType.INVALID_CLEAR_PIN_COUNT);
         }
 
         this.pin = pin;
+    }
+
+    static Pin of(int pinNumber) {
+        return new Pin(pinNumber);
     }
 }
